pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medisync"
version = "1.0.0"
authors = [
    { name = "MediSync Team" }
]
description = "Emergency Medical Services Portal"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Flask>=2.3.3",
    "Flask-SQLAlchemy>=3.1.1",
    "Flask-Limiter>=3.5.0",
    "Flask-Caching>=2.1.0",
    "Flask-WTF>=1.2.1",
    "marshmallow>=3.20.1",
    "cryptography>=41.0.4",
    "python-dotenv>=1.0.0",
    "redis>=5.0.1",
    "gunicorn>=21.2.0",
    "SQLAlchemy>=2.0.22",
    "Werkzeug>=2.3.7",
    "psycopg2-binary>=2.9.9",
    "python-jose>=3.3.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/medisync"
"Bug Tracker" = "https://github.com/yourusername/medisync/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["api"] 